wip(src/ra/algo/phase2): group forming, calc group degree, zone allocation, recode unallocated ue
src/ra/algo/phase2: write the algo of phase2 for gNB and eNB including
calculate the required layer to use,
forming groups and calculate the residual degree for gNB,
allocate the zone group on to the frame of gNB,
and allocate a single zone on the frame of gNB or eNB.
But ignore the cp value for eNB zone allocation.
I'm thinking, will not mentioning the cp value be easier to understand?

tests/proof_of_concept: testing phase 1 and phase 2 for gNB then eNB

src/ra/ds/frame:
* remove the list of RB in layer because there's no usage of the list.
We simply need to know two messages about RB.
One is if (i,j) is allocated, this could be found in layer.BU array.
The other one is which RBs are the UE using, this is saved in UE.rb.
* refactor allocate_resource_block.
Refactor the assertion for "the RB is in the range of the frame".
And append the new RB to the UE.
* Add a new property for returning the current available offset of a layer
for the usage of zone allocation.

src/ra/ds/zone:
* _Bin: Add a new property to return the remaining spaces of a bin.
To check of a zone can be appended.
* zoneGroup: Remove the method append_zone
because this method will not only be used in zoneGroup
but also the direct allocation to the layer in the algorithm
such as the eNB zone allocation and the gNB zone allocation after zone group allocation.