fix(src/ra/algo/phase2): nodeb type checking error feat(src/ra/algo/phase2): zone_group_allocation fill the frame as full as possible wip(src/ra/algo/phase2): refactor zone allocation
fix: nodeb type checking error
The type of self.nodeb is the class no the enum.

feat: zone_group_allocation fill the frame as full as possible
Fill up the frame with zone groups according to the degree.
If the remaing space isn't enough for higher degree, allocate a lower one that fits.
This modification is to fill the frame with zones to minimize the INI.

wip: refactor zone allocation
* Move the method from phase 2 to ds/frame.
To simplify the code and improve the readability.
* Return 'True' if zone is succefully allocated.
To unite the code logic.